+++ Model/Koszt.cs
@@ -13,9 +13,9 @@ namespace Ksiegowosc.Model
         public string NumerDokumentu { get; set; }
         public DateTime DataWystawienia { get; set; }
         public string Opis { get; set; }
-        public int KwotaNetto { get; set; }
-        public int StawkaVat { get; set; }
-        public int KwotaBrutto { get; set; }
+        public decimal KwotaNetto { get; set; }
+        public decimal StawkaVat { get; set; }
+        public decimal KwotaBrutto { get; set; }
         public string Kategoria { get; set; }
 
         public Kontrahent Kontrahent { get; set; }
